docker/data-quality/npi/requirements.txt
great-expectations==0.18.0
pandas==2.2.0
pyarrow==15.0.0
azure-storage-file-datalake==12.14.0
azure-identity==1.15.0
